feat: Add rank filter (#130)
* feat: Added enhance contrast command

* feat: Added rank filter

* chore: Fixed enhance contrast

* feat: Allow to create tags

* chore: Fixed contrast

* chore: Fixed image cache settings

* feat: Generate preview using multiple threads

* chore: Allow higher contrast

* fix: Image cache was displayed wrong in pipeline flow

* fix: Fixed auto enhancement

* feat: Add clear button to history

+++ src/backend/commands/image_functions/edge_detection_canny/edge_detection_canny_settings.hpp
@@ -19,12 +19,12 @@ public:
   //
   // Lower threshold for canny edge detection in range [0-255]
   //
-  uint16_t thresholdMin = 10;
+  float thresholdMin = 10;
 
   //
   // Upper threshold for canny edge detection in range [0-255]
   //
-  uint16_t thresholdMax = 150;
+  float thresholdMax = 150;
 
   /////////////////////////////////////////////////////
   void check() const
